Ts files to Js
